+++ src/quo/_spinners.py
@@ -22,14 +22,12 @@ MIT License
     ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGs IN THE SOFTWARE.
 
     """
+    # This code has been reformatted...
 
-    # This code has been reformatted to suit our purpose...
-
-class SPINNERS():
-    arrows3 = r"▹▹▹▹▹▸▹▹▹▹ ▹▸▹▹▹ ▹▹▸▹▹ ▹▹▹▸▹ ▹▹▹▹▸"
+class Spinners:
+    arrows = r"▹▹▹▹▹▸▹▹▹▹ ▹▸▹▹▹ ▹▹▸▹▹ ▹▹▹▸▹ ▹▹▹▹▸"
     dots = r"⠋⠙⠹⠸⠼⠴⠦⠧⠇⠏"
     dots2 = r"⣾⣽⣻⢿⡿⣟⣯⣷"
     dots3 = r"⠋⠙⠚⠞⠖⠦⠴⠲⠳⠓"
     dots4 = r"⠄⠆⠇⠋⠙⠸⠰⠠⠰⠸⠙⠋⠇⠆"
-
     hamburger = r"☱☲☴"