+++ src/quo/rule.py
@@ -1,6 +1,6 @@
 class Rule:
-    def __init__(self, height:int = 1, char: str="\u2501"):
-        self.height= height
+    def __init__(self, lines:int=1, char: str="\u2501"):
+        self.lines = lines
         self.char = char
 
        
@@ -22,17 +22,19 @@ class Rule:
             for i in range(width):
                 # Get the current color from the list
                 selectedColor = colorList[colorIndex]
-                # Print the current character in the current color
-                print(colored(self.char, selectedColor), end='')
-                # Increment the color index
-                colorIndex = (colorIndex + 1) % len(colorList)
-    
+
+                for f in range(self.lines):
+                    # Print the current character in the current color
+                    print(colored(self.char, selectedColor), end='')
+                    # Increment the color index
+                    colorIndex = (colorIndex + 1) % len(colorList)
+                    
             # Print a newline to move to the next line
             print()
 
         else:
             from quo.layout.containers import Window
             from quo.shortcuts.utils import container
-            container(Window(char=self.char, height=self.height, style="fg:" +color),bind=False)
+            container(Window(char=self.char, height=self.lines, style="fg:" +color),bind=False)